+++ keyboards/helix/rev3_5rows/rules.mk
@@ -8,3 +8,4 @@ DIP_SWITCH_ENABLE = yes
 LTO_ENABLE = yes
 
 SRC += oled_display.c
+SRC += oled_layer_name.c
